Fix travis error
